Default to CtrlPMixed
Without reading the documentation, it's not obvious that CtrlP has three modes. Even knowing that there were modes, I did not learn how to switch between them until I stumbled upon it accidentally after about a month. I only now realized that there was a mixed mode, which is really *exactly what I want*. 

Mixed mode is similar to the behavior of TextMate and SublimeText, which is pretty intuitive. MRU and buffer files are very frequently reopened, which makes Mixed mode great.

With the occasional difficulty of finding some files because of #110, having Mixed mode on by default will at least allieve this issue after the first time the file is opened.

Is there any reason I'm not thinking of why Mixed mode should not be the default?