Make README screenshot narrower to fit
Make README screenshot narrower so it fits on the GitHub project home page

The current screenshot as displayed on https://github.com/kien/ctrlp.vim is squished to fit within 722px, introducing aliasing and making the screenshot hard to read. This updated image is 720px, so it will fit, and provide a better first impression of ctrlp.vim to people visiting the project page.

I produced this screenshot in Photoshop by effectively deleting some of the empty space in the middle. The window in the image might not be a realistically-sized one – its width might not be a multiple of the character cell width – but nobody will notice.