Fix auto selection of tag from tselect menu for tags in the same file.
Starting with commit 83397bd0, when possible Ctrlp tries to autoselect
a tag from from the `tselect` menu.

This has a significant flaw though, as it does not select the correct tag
when there are multiple tags with the same name in the same file, instead
preferring to always use the last tag from that file.

By matching not just on filename of the selected tag, but also on the
tagaddress (generally a regex or linenumber), we ensure that the correct
tag is always selected from the tselect prompt.