Separate path correctly for *sh on Windows
In user command, `%s` is replaced with a target directory path. On
Windows, this path is always separated by backslashes. If a user
configures `shell` option to `sh` (or something similar), a shell cannot
understand a target directory path and probably cannot list files.

This commit checks the `shell` option on Windows. If this option
contains `sh`, a target directory path is separated by forward slashes.