Call ctags with --file-scope=yes argument
Universal Ctags requires the `F` flag to be present in the --extra
argument in order to report file-scoped tags. The easiest way to make an
argument list compatible with both Exuberant Ctags and Universal Ctags
is to pass the --file-scope=yes argument in both cases. This will cause
Universal Ctags to report file-scoped tags, and be a no-op on Exuberant
Ctags since it does that already even with an empty --extra argument.