Add g:ctrlp_regex_always_higlight option
CtrlP highlights only path and files as you type pattern.  Other
extensions, such as 'lines' or 'tags' do not have highlighting as you
type search pattern.  With this patch other other extensions also can
have pattern highlighting. For this, it is necessary to add:

    let g:ctrlp_regex_always_higlight = 1

This patch works only when: g:ctrlp_regexp = 1. Which is quite safe as
universal approach.

Signed-off-by: Andrei Stepanov <astepano@redhat.com>